Redo Strife TITLEPIC and RGELOGO
Changed my mind about these. I'll handle these the same way I did the Heretic title screen (expand, and then fill with the exact empty space colour).

For the TITLEPIC, an edit to the original 4:3 is still needed because there appears to be some stray anti-aliasing artefacts on the top-left border of the image